+++ frontend_react/react_src/src/api-state/apiClient.ts
@@ -1,14 +1,32 @@
-import { getBaseUrl } from "@/utils";
-import axios from "axios";
+import { getBaseUrl, isProd } from "@/utils";
+import axios, { CreateAxiosDefaults } from "axios";
 
-const apiClient = axios.create({
-  baseURL: `${getBaseUrl()}/api`,
+type ConfigDefaults = {
+  baseURL: string;
   headers: {
-    Accept: "application/json",
-    "Content-type": "application/json",
-  },
-  withCredentials: true,
-});
+    Accept: string;
+    "Content-type": string;
+    "x-user-email"?: string;
+  };
+  withCredentials: boolean;
+};
+
+const getAxiosConfig = (): CreateAxiosDefaults<ConfigDefaults> => {
+  const config: ConfigDefaults = {
+    baseURL: `${getBaseUrl()}/api`,
+    headers: {
+      Accept: "application/json",
+      "Content-type": "application/json",
+    },
+    withCredentials: true,
+  };
+  if (!isProd()) {
+    config.headers["x-user-email"] = "user@domain.com";
+  }
+  return config;
+};
+
+const apiClient = axios.create(getAxiosConfig());
 
 export default apiClient;