+++ frontend_react/react_src/src/utils.ts
@@ -1,6 +1,10 @@
+
+export function isProd() {
+  return import.meta.env.MODE === "production";
+}
 export function getBaseUrl() {
   let baseUrl = "http://localhost:8080";
-  if (import.meta.env.MODE === "production") {
+  if (isProd()) {
     // expected valid url https://app.datarobot.com/custom_applications/{appId}/api
     const fullUrl = window.location.origin + window.location.pathname;
     baseUrl = fullUrl.split("/").splice(0, 5).join("/");