+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Improved error handling when prompting (react)
+
 ## [0.1.13] - 2025-05-06
 
 ### Fixed